# api/connect-service-openapi.yaml
openapi: 3.0.3
info:
  title: Connect Service API
  description: |
    A robust Spring Boot reactive microservice for orchestrating data flow between multiple external services.
    This service acts as a connector hub, processing XML/JSON payloads through configurable pipelines
    and managing authentication with various external systems including Apigee Gateway and Fenergo.
    
    ## Key Features
    - **Reactive Processing**: Built with Spring WebFlux and Project Reactor
    - **Multi-Service Orchestration**: Connect and orchestrate multiple external services
    - **Flexible Authentication**: Support for Apigee Gateway, OAuth2, and mTLS
    - **Audit & Monitoring**: Complete request tracking with correlation IDs
    - **SSL MongoDB**: Advanced MongoDB connectivity with SSL/TLS
    - **Vault Integration**: Secure credential management with HashiCorp Vault
    
    ## Authentication
    All requests must be authenticated via Apigee Gateway. The service validates incoming requests
    and manages OAuth2 tokens for outbound calls to external services.
    
    ## Request Processing Flow
    1. Receive XML/JSON payload with request type enum
    2. Convert XML to JSON (if needed)
    3. Call external service based on request type
    4. Call Fenergo API with processed data
    5. Store all data in MongoDB with audit trail
    
    ## Monitoring
    - Health checks available at `/connect/health`
    - Metrics exposed at `/actuator/prometheus`
...
  version: 1.0.0
  contact:
...
servers:
...
security:
  - ApigeeAuth: []
paths:
...
  /v1/connect/status/{requestId}:
    get:
      tags:
        - Connect
      summary: Get request processing status
      description: |
        Retrieve the current status and details of a processing request.
        This endpoint provides real-time information about the request processing
        including current step, completion status, and any error details.
      operationId: getRequestStatus
      parameters:
        - name: requestId
          in: path
          required: true
          description: Unique identifier of the processing request
          schema:
            type: string
            example: "req_123456789"
      responses:
        '200':
          description: Request status retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestStatusResponse'
              examples:
                processing:
                  summary: Request Processing
                  description: Request is currently being processed
                  value:
                    requestId: "req_123456789"
                    status: "PROCESSING"
                    currentStep: "EXTERNAL_API_CALL"
                    progress: 60
                    message: "Processing external API call"
                    correlationId: "550e8400-e29b-41d4-a716-446655440000"
                    createdAt: "2024-01-15T10:30:00Z"
                    updatedAt: "2024-01-15T10:32:00Z"
                    steps:
                      - step: "XML_TO_JSON_CONVERSION"
                        status: "COMPLETED"
                        completedAt: "2024-01-15T10:30:30Z"
                        duration: 500
                      - step: "EXTERNAL_API_CALL"
                        status: "IN_PROGRESS"
                        startedAt: "2024-01-15T10:30:30Z"
                      - step: "FENERGO_API_CALL"
                        status: "PENDING"
                      - step: "COMPLETION"
                        status: "PENDING"
                completed:
                  summary: Request Completed
                  description: Request has been completed successfully
                  value:
                    requestId: "req_123456789"
                    status: "COMPLETED"
                    currentStep: "COMPLETION"
                    progress: 100
                    message: "Request completed successfully"
                    correlationId: "550e8400-e29b-41d4-a716-446655440000"
                    createdAt: "2024-01-15T10:30:00Z"
                    updatedAt: "2024-01-15T10:35:00Z"
                    completedAt: "2024-01-15T10:35:00Z"
                    totalDuration: 300000
                    steps:
                      - step: "XML_TO_JSON_CONVERSION"
                        status: "COMPLETED"
                        completedAt: "2024-01-15T10:30:30Z"
                        duration: 500
                      - step: "EXTERNAL_API_CALL"
                        status: "COMPLETED"
                        completedAt: "2024-01-15T10:32:00Z"
                        duration: 90000
                      - step: "FENERGO_API_CALL"
                        status: "COMPLETED"
                        completedAt: "2024-01-15T10:34:00Z"
                        duration: 120000
                      - step: "COMPLETION"
                        status: "COMPLETED"
                        completedAt: "2024-01-15T10:35:00Z"
                        duration: 1000
                failed:
                  summary: Request Failed
                  description: Request processing failed
                  value:
                    requestId: "req_123456789"
                    status: "FAILED"
                    currentStep: "EXTERNAL_API_CALL"
                    progress: 40
                    message: "External API call failed"
                    correlationId: "550e8400-e29b-41d4-a716-446655440000"
                    createdAt: "2024-01-15T10:30:00Z"
                    updatedAt: "2024-01-15T10:32:00Z"
                    error:
                      code: "EXTERNAL_API_ERROR"
                      message: "Connection timeout to external service"
                      details: "Service unavailable: timeout after 30 seconds"
                    steps:
                      - step: "XML_TO_JSON_CONVERSION"
                        status: "COMPLETED"
                        completedAt: "2024-01-15T10:30:30Z"
                        duration: 500
                      - step: "EXTERNAL_API_CALL"
                        status: "FAILED"
                        failedAt: "2024-01-15T10:32:00Z"
                        error: "Connection timeout"
                      - step: "FENERGO_API_CALL"
                        status: "CANCELLED"
                      - step: "COMPLETION"
                        status: "CANCELLED"
        '404':
          description: Request not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                not_found:
                  summary: Request Not Found
                  description: The specified request ID was not found
                  value:
                    error: "REQUEST_NOT_FOUND"
                    message: "Request with ID 'req_123456789' not found"
                    details: "Please verify the request ID and try again"
                    correlationId: "550e8400-e29b-41d4-a716-446655440000"
                    timestamp: "2024-01-15T10:30:00Z"
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    ApigeeAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Authentication via Apigee Gateway using JWT tokens.
        Include the Apigee token in the Authorization header as: `Bearer <token>`
  schemas:
...
    RequestStatusResponse:
      type: object
      required:
        - requestId
        - status
        - correlationId
        - createdAt
        - updatedAt
      properties:
        requestId:
          type: string
          description: Unique identifier for the processing request
          example: "req_123456789"
        status:
          $ref: '#/components/schemas/ProcessingStatus'
        currentStep:
          type: string
          description: Current processing step
          example: "EXTERNAL_API_CALL"
        progress:
          type: integer
          minimum: 0
          maximum: 100
          description: Processing progress percentage
          example: 60
        message:
          type: string
          description: Current status message
          example: "Processing external API call"
        correlationId:
          type: string
          description: Correlation ID for tracking
          example: "550e8400-e29b-41d4-a716-446655440000"
        createdAt:
          type: string
          format: date-time
          description: Request creation timestamp
          example: "2024-01-15T10:30:00Z"
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
          example: "2024-01-15T10:32:00Z"
        completedAt:
          type: string
          format: date-time
          description: Completion timestamp (if completed)
          example: "2024-01-15T10:35:00Z"
        totalDuration:
          type: integer
          description: Total processing duration in milliseconds
          example: 300000
        error:
          $ref: '#/components/schemas/ErrorDetails'
        steps:
          type: array
          description: Processing steps with their status
          items:
            $ref: '#/components/schemas/ProcessingStep'
    ProcessingStep:
      type: object
      required:
        - step
        - status
      properties:
        step:
          type: string
          description: Step name
          example: "XML_TO_JSON_CONVERSION"
        status:
          type: string
          enum: [PENDING, IN_PROGRESS, COMPLETED, FAILED, CANCELLED]
          description: Step status
          example: "COMPLETED"
        startedAt:
          type: string
          format: date-time
          description: Step start timestamp
          example: "2024-01-15T10:30:00Z"
        completedAt:
          type: string
          format: date-time
          description: Step completion timestamp
          example: "2024-01-15T10:30:30Z"
        failedAt:
          type: string
          format: date-time
          description: Step failure timestamp
          example: "2024-01-15T10:32:00Z"
        duration:
          type: integer
          description: Step duration in milliseconds
          example: 500
        error:
          type: string
          description: Error message if step failed
          example: "Connection timeout"
...
    ProcessingStatus:
      type: string
      enum:
        - PENDING
        - PROCESSING
        - COMPLETED
        - FAILED
        - CANCELLED
      description: |
        Current processing status:
        - **PENDING**: Request is queued for processing
        - **PROCESSING**: Request is currently being processed
        - **COMPLETED**: Request has been completed successfully
        - **FAILED**: Request processing failed
        - **CANCELLED**: Request processing was cancelled
      example: "PROCESSING"
    ErrorResponse:
      type: object
      required:
        - error
        - message
        - correlationId
        - timestamp
      properties:
        error:
          type: string
          description: Error code
          example: "VALIDATION_ERROR"
        message:
          type: string
          description: Error message
          example: "Invalid request type: UNSUPPORTED_TYPE"
        details:
          type: string
          description: Additional error details
          example: "Request type must be one of: CUSTOMER_ONBOARDING, DOCUMENT_VERIFICATION, RISK_ASSESSMENT, COMPLIANCE_CHECK, GENERIC"
        correlationId:
          type: string
          description: Correlation ID for tracking
          example: "550e8400-e29b-41d4-a716-446655440000"
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
          example: "2024-01-15T10:30:00Z"
    ErrorDetails:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Error code
          example: "EXTERNAL_API_ERROR"
        message:
          type: string
          description: Error message
          example: "Connection timeout to external service"
        details:
          type: string
          description: Additional error details
          example: "Service unavailable: timeout after 30 seconds"
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
          example: "2024-01-15T10:32:00Z"
...